+++ apps/community/src/components/lab/lab-card.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
-import type { Lab } from '~/app/lab/remotes';
 import * as styles from '~/components/lab/lab-card.css';
+import type { Lab } from '~/types/lab';
 
 function LabCard({ lab }: { lab: Lab }) {
   return (
